Refactor: Migrate to ESP-IDF and improve BLE functionality
This commit refactors the project to use the ESP-IDF framework instead of Arduino. It includes significant improvements to the BLE Manager, including the implementation of GATT server events, advertising, and notifications. The LED Controller and Power Manager have also been updated to use ESP-IDF APIs. Additionally, OTA functionality has been enhanced, and a new TimeUtils header provides millisecond and delay functions.

+++ src/PowerManager.h
@@ -1,14 +1,17 @@
 #ifndef POWER_MANAGER_H
 #define POWER_MANAGER_H
 
-#include <Arduino.h>
 #include <esp_adc_cal.h>
 #include <esp_sleep.h>
 #include <esp_task_wdt.h>
 #include <esp_ota_ops.h>
+#include <freertos/FreeRTOS.h>
+#include <freertos/task.h>
+#include <driver/gpio.h>
 #include "Config.h"
+#include "TimeUtils.h"
 
-class LEDController; // Forward declaration
+class LEDController;
 
 class PowerManager 
 {
@@ -16,45 +19,28 @@ private:
     LEDController* ledController;
     esp_adc_cal_characteristics_t adcChars;
     TaskHandle_t batteryTaskHandle;
-    unsigned long lastDisconnectTime;
+    uint32_t lastDisconnectTime;
     bool isGoingToSleep;
     bool justWokenUp;
     bool isInitialized;
-    
-    // Callback function pointer for battery level updates
     void (*batteryCallback)(int percentage);
-    
 public:
     PowerManager(LEDController* leds);
-    
-    // Initialization
     bool begin();
     void setLedController(LEDController* leds) { ledController = leds; }
-    
-    // Battery monitoring
     int getBatteryPercentage();
     void setBatteryCallback(void (*callback)(int));
     void startBatteryMonitoringTask();
     void stopBatteryMonitoringTask();
-    
-    // Sleep/Wake functionality
     void goToDeepSleep(bool skipLedBlink = false);
     bool validateWakeUp();
     bool isJustWokenUp() const { return justWokenUp; }
     void clearWakeUpFlag() { justWokenUp = false; }
-    
-    // Connection state management
     void setConnected(bool connected);
     bool shouldAutoSleep();
-    
-    // Charging detection
     bool isCharging();
-    
-    // Button handling
     bool isButtonPressed();
-    
-    // Task function (static)
     static void batteryMonitorTaskFunction(void* pvParameters);
 };
 
-#endif // POWER_MANAGER_H
+#endif
